+++ commanderbot/ext/jira/jira_cog.py
@@ -4,6 +4,7 @@ from discord import Embed, Interaction, ui
 from discord.app_commands import (
     Transform,
     Transformer,
+    allowed_contexts,
     allowed_installs,
     command,
     describe,
@@ -50,6 +51,7 @@ class JiraCog(Cog, name="commanderbot.ext.jira"):
     @command(name="jira", description="Query an issue from Jira")
     @describe(query="The issue ID or URL to query")
     @allowed_installs(guilds=True, users=True)
+    @allowed_contexts(guilds=True, dms=True, private_channels=True)
     async def cmd_jira(
         self,
         interaction: Interaction,