+++ CHANGELOG.md
@@ -8,6 +8,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+### Changed
+
+## [0.21.0]
+
+### Added
+
+- Added a context menu command for kicking compromised users
+
+### Changed
+
+- Updated Python to 3.14
+- Updated discord.py to 2.6
+- Fixed an issue where the Bedrock changelog feed wasn't finding the thumbnail
+- Updated the version regex the Bedrock changelog feed was using to support the new version format
+
 ## [0.20.0] - 2025-05-03
 
 ### Added
@@ -382,7 +397,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Implemented `status` and `faq` extensions as an exercise for developing `commanderbot-lib`
 
-[unreleased]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.20.0...HEAD
+[unreleased]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/MinecraftCommands/commanderbot-py/compare/v0.17.0...v0.18.0